fix: address code review findings from Postgres migration
Critical fixes:
- Remove get_metadata/set_metadata methods that query non-existent _sync_metadata table
- Fix connection leak in ConnectionProvider when execute fails after connect

Major fixes:
- Remove sync_metadata table creation from admin client initialize_schema()
- Remove deprecated CREATE_SYNC_METADATA_TABLE constants from both schema files
- Remove DEFAULT_SYNC_METADATA constant

Simplifications:
- Consolidate duplicate check_database_connection functions into db.connection module
- Consolidate duplicate _to_str helper functions into new db.helpers module
- Remove dead SyncError exception classes from both admin and app clients
- Remove dead INTEGRITY_CHECK_QUERY and FOREIGN_KEYS_QUERY constants

Documentation:
- Update README.md to replace all Turso references with PostgreSQL
- Update AGENTS.md and CLAUDE.md to fix Turso reference

Tests:
- Update test files to patch correct module path for check_database_connection
- Remove SyncError import from skipped test file
