+++ src/stream_of_worship/db/postgres_schema.py
@@ -18,7 +18,6 @@ from stream_of_worship.admin.db.schema import (
     CREATE_SONGS_TABLE,
     CREATE_SONGS_UPDATE_TRIGGER,
     CREATE_UPDATE_TIMESTAMP_FUNCTION,
-    DEFAULT_SYNC_METADATA,
     RECORDING_COLUMNS_FOR_JOIN,
     RECORDING_COLUMN_COUNT,
     ROW_COUNT_QUERY,
@@ -63,7 +62,6 @@ __all__ = [
     "CREATE_SONGS_TABLE",
     "CREATE_SONGS_UPDATE_TRIGGER",
     "CREATE_UPDATE_TIMESTAMP_FUNCTION",
-    "DEFAULT_SYNC_METADATA",
     "RECORDING_COLUMNS_FOR_JOIN",
     "RECORDING_COLUMN_COUNT",
     "ROW_COUNT_QUERY",
